+++ src/workbench/flat.ts
@@ -484,6 +484,8 @@ export function flatWorkbench(
     "settings.dropdownBackground": `${palette.bg}`,
     "settings.dropdownForeground": `${palette.aqua}`,
     "settings.dropdownBorder": `${palette.bg7}`,
+    "settings.modifiedItemIndicator": `${palette.grey0}`,
+    "settings.focusedRowBackground": `${palette.bg2}`,
     "editorLightBulb.foreground": `${palette.yellow}`,
     "editorLightBulbAutoFix.foreground": `${palette.aqua}`,
     "welcomePage.buttonBackground": `${palette.bg2}`,